+++ inst/dust/model.cpp
@@ -52,6 +52,7 @@ __host__ __device__ T odin_max(T x, T y) {
 // [[dust::param(p_I, has_default = FALSE, default_value = NULL, rank = 0, min = -Inf, max = Inf, integer = FALSE)]]
 // [[dust::param(p_R, has_default = FALSE, default_value = NULL, rank = 0, min = -Inf, max = Inf, integer = FALSE)]]
 // [[dust::param(p_S, has_default = FALSE, default_value = NULL, rank = 0, min = -Inf, max = Inf, integer = FALSE)]]
+// [[dust::param(q_F, has_default = FALSE, default_value = NULL, rank = 1, min = -Inf, max = Inf, integer = FALSE)]]
 // [[dust::param(sigma, has_default = FALSE, default_value = NULL, rank = 0, min = -Inf, max = Inf, integer = FALSE)]]
 // [[dust::param(t_s, has_default = FALSE, default_value = NULL, rank = 0, min = -Inf, max = Inf, integer = FALSE)]]
 // [[dust::param(theta_A, has_default = FALSE, default_value = NULL, rank = 0, min = -Inf, max = Inf, integer = FALSE)]]
@@ -239,6 +240,7 @@ public:
     int dim_omega;
     int dim_prev_A;
     int dim_prev_R;
+    int dim_q_F;
     int dim_r_A;
     int dim_r_E;
     int dim_r_F;
@@ -267,6 +269,7 @@ public:
     real_type initial_daily_gas_pharyngitis_rate_45_64;
     real_type initial_daily_gas_pharyngitis_rate_65_74;
     real_type initial_daily_gas_pharyngitis_rate_75;
+    real_type initial_daily_scarlet_fever_rate;
     real_type initial_daily_scarlet_fever_rate_04;
     real_type initial_daily_scarlet_fever_rate_05_14;
     real_type initial_daily_scarlet_fever_rate_15_44;
@@ -279,6 +282,7 @@ public:
     real_type initial_net_leavers_inc;
     std::vector<real_type> initial_prev_A;
     std::vector<real_type> initial_prev_R;
+    real_type initial_scarlet_fever_cases;
     real_type initial_scarlet_fever_inc;
     real_type initial_scarlet_fever_inc_04;
     real_type initial_scarlet_fever_inc_05_14;
@@ -310,6 +314,7 @@ public:
     real_type p_R;
     real_type p_S;
     real_type pi;
+    std::vector<real_type> q_F;
     std::vector<real_type> r_A;
     std::vector<real_type> r_E;
     std::vector<real_type> r_F;
@@ -385,38 +390,40 @@ public:
     shared(pars.shared), internal(pars.internal) {
   }
   size_t size() {
-    return shared->dim_A + shared->dim_E + shared->dim_F + shared->dim_N + shared->dim_P + shared->dim_R + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
+    return shared->dim_A + shared->dim_E + shared->dim_F + shared->dim_N + shared->dim_P + shared->dim_R + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
   }
   std::vector<real_type> initial(size_t step) {
-    std::vector<real_type> state(shared->dim_A + shared->dim_E + shared->dim_F + shared->dim_N + shared->dim_P + shared->dim_R + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27);
+    std::vector<real_type> state(shared->dim_A + shared->dim_E + shared->dim_F + shared->dim_N + shared->dim_P + shared->dim_R + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29);
     state[0] = shared->initial_time;
     state[1] = shared->initial_infections_inc;
     state[2] = shared->initial_gas_pharyngitis_inc;
     state[3] = shared->initial_scarlet_fever_inc;
-    state[4] = shared->initial_igas_inc;
-    state[5] = shared->initial_births_inc;
-    state[6] = shared->initial_net_leavers_inc;
-    state[7] = shared->initial_beta_t;
-    state[8] = shared->initial_daily_gas_pharyngitis_rate;
-    state[9] = shared->initial_daily_gas_pharyngitis_rate_04;
-    state[10] = shared->initial_daily_gas_pharyngitis_rate_05_14;
-    state[11] = shared->initial_daily_gas_pharyngitis_rate_15_44;
-    state[12] = shared->initial_daily_gas_pharyngitis_rate_45_64;
-    state[13] = shared->initial_daily_gas_pharyngitis_rate_65_74;
-    state[14] = shared->initial_daily_gas_pharyngitis_rate_75;
-    state[15] = shared->initial_daily_scarlet_fever_rate_04;
-    state[16] = shared->initial_daily_scarlet_fever_rate_05_14;
-    state[17] = shared->initial_daily_scarlet_fever_rate_15_44;
-    state[18] = shared->initial_daily_scarlet_fever_rate_45_64;
-    state[19] = shared->initial_daily_scarlet_fever_rate_65_74;
-    state[20] = shared->initial_daily_scarlet_fever_rate_75;
-    state[21] = shared->initial_scarlet_fever_inc_04;
-    state[22] = shared->initial_scarlet_fever_inc_05_14;
-    state[23] = shared->initial_scarlet_fever_inc_15_44;
-    state[24] = shared->initial_scarlet_fever_inc_45_64;
-    state[25] = shared->initial_scarlet_fever_inc_65_74;
-    state[26] = shared->initial_scarlet_fever_inc_75;
-    std::copy(shared->initial_U.begin(), shared->initial_U.end(), state.begin() + 27);
+    state[4] = shared->initial_scarlet_fever_cases;
+    state[5] = shared->initial_igas_inc;
+    state[6] = shared->initial_births_inc;
+    state[7] = shared->initial_net_leavers_inc;
+    state[8] = shared->initial_beta_t;
+    state[9] = shared->initial_daily_gas_pharyngitis_rate;
+    state[10] = shared->initial_daily_scarlet_fever_rate;
+    state[11] = shared->initial_daily_gas_pharyngitis_rate_04;
+    state[12] = shared->initial_daily_gas_pharyngitis_rate_05_14;
+    state[13] = shared->initial_daily_gas_pharyngitis_rate_15_44;
+    state[14] = shared->initial_daily_gas_pharyngitis_rate_45_64;
+    state[15] = shared->initial_daily_gas_pharyngitis_rate_65_74;
+    state[16] = shared->initial_daily_gas_pharyngitis_rate_75;
+    state[17] = shared->initial_daily_scarlet_fever_rate_04;
+    state[18] = shared->initial_daily_scarlet_fever_rate_05_14;
+    state[19] = shared->initial_daily_scarlet_fever_rate_15_44;
+    state[20] = shared->initial_daily_scarlet_fever_rate_45_64;
+    state[21] = shared->initial_daily_scarlet_fever_rate_65_74;
+    state[22] = shared->initial_daily_scarlet_fever_rate_75;
+    state[23] = shared->initial_scarlet_fever_inc_04;
+    state[24] = shared->initial_scarlet_fever_inc_05_14;
+    state[25] = shared->initial_scarlet_fever_inc_15_44;
+    state[26] = shared->initial_scarlet_fever_inc_45_64;
+    state[27] = shared->initial_scarlet_fever_inc_65_74;
+    state[28] = shared->initial_scarlet_fever_inc_75;
+    std::copy(shared->initial_U.begin(), shared->initial_U.end(), state.begin() + 29);
     std::copy(shared->initial_N.begin(), shared->initial_N.end(), state.begin() + shared->offset_variable_N);
     std::copy(shared->initial_prev_A.begin(), shared->initial_prev_A.end(), state.begin() + shared->offset_variable_prev_A);
     std::copy(shared->initial_prev_R.begin(), shared->initial_prev_R.end(), state.begin() + shared->offset_variable_prev_R);
@@ -429,7 +436,7 @@ public:
     return state;
   }
   void update(size_t step, const real_type * state, rng_state_type& rng_state, real_type * state_next) {
-    const real_type * U = state + 27;
+    const real_type * U = state + 29;
     const real_type * A = state + shared->offset_variable_A;
     const real_type * E = state + shared->offset_variable_E;
     const real_type * S = state + shared->offset_variable_S;
@@ -437,13 +444,14 @@ public:
     const real_type * F = state + shared->offset_variable_F;
     const real_type * R = state + shared->offset_variable_R;
     const real_type * N = state + shared->offset_variable_N;
-    const real_type births_inc = state[5];
-    const real_type net_leavers_inc = state[6];
+    const real_type births_inc = state[6];
+    const real_type net_leavers_inc = state[7];
     const real_type infections_inc = state[1];
-    const real_type igas_inc = state[4];
-    const real_type beta_t = state[7];
-    state_next[7] = shared->beta * (1 + shared->sigma * std::cos(2 * shared->pi * (shared->t0 + step - shared->t_s) / (real_type) 365.25));
+    const real_type igas_inc = state[5];
+    const real_type beta_t = state[8];
+    state_next[8] = shared->beta * (1 + shared->sigma * std::cos(2 * shared->pi * (shared->t0 + step - shared->t_s) / (real_type) 365.25));
     real_type alpha_t = (static_cast<int>(step) >= shared->dim_alpha ? shared->alpha[shared->dim_alpha - 1] : shared->alpha[step + 1 - 1]);
+    real_type q_F_t = (static_cast<int>(step) >= shared->dim_q_F ? shared->q_F[shared->dim_q_F - 1] : shared->q_F[step + 1 - 1]);
     state_next[0] = (step + 1) * shared->dt;
     for (int i = 1; i <= shared->dim_n_Ui; ++i) {
       internal.n_Ui[i - 1] = ((i > 1 ? U[i - 1 - 1] : 0)) - ((i < shared->n_group ? U[i - 1] : 0));
@@ -532,7 +540,7 @@ public:
         internal.n_Sx[i - 1 + shared->dim_n_Sx_1 * (j - 1)] = std::round(S[shared->dim_S_1 * (j - 1) + i - 1] * shared->omega[i - 1] * shared->dt);
       }
     }
-    state_next[5] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_xU.data(), 0, shared->dim_n_xU) : births_inc + odin_sum1<real_type>(internal.n_xU.data(), 0, shared->dim_n_xU)));
+    state_next[6] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_xU.data(), 0, shared->dim_n_xU) : births_inc + odin_sum1<real_type>(internal.n_xU.data(), 0, shared->dim_n_xU)));
     for (int i = 1; i <= shared->dim_dem_A_1; ++i) {
       for (int j = 1; j <= shared->dim_dem_A_2; ++j) {
         internal.dem_A[i - 1 + shared->dim_dem_A_1 * (j - 1)] = std::round(internal.n_Ai[shared->dim_n_Ai_1 * (j - 1) + i - 1] * shared->r_age * shared->dt) - internal.n_Ax[shared->dim_n_Ax_1 * (j - 1) + i - 1];
@@ -608,7 +616,7 @@ public:
     for (int i = 1; i <= shared->dim_r_U; ++i) {
       internal.r_U[i - 1] = internal.foi[i - 1];
     }
-    state_next[6] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_Nx.data(), 0, shared->dim_n_Nx) : net_leavers_inc + odin_sum1<real_type>(internal.n_Nx.data(), 0, shared->dim_n_Nx)));
+    state_next[7] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_Nx.data(), 0, shared->dim_n_Nx) : net_leavers_inc + odin_sum1<real_type>(internal.n_Nx.data(), 0, shared->dim_n_Nx)));
     for (int i = 1; i <= shared->dim_n_AR; ++i) {
       internal.n_AR[i - 1] = dust::random::binomial<real_type>(rng_state, internal.n_A[shared->dim_n_A_1 * (shared->k_A - 1) + i - 1], shared->p_R);
     }
@@ -663,7 +671,7 @@ public:
     for (int i = 1; i <= shared->dim_scarlet_fever_inc_by_group; ++i) {
       internal.scarlet_fever_inc_by_group[i - 1] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? internal.n_PF[i - 1] : internal.scarlet_fever_inc_by_group[i - 1] + internal.n_PF[i - 1]));
     }
-    state_next[4] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_I.data(), 0, shared->dim_n_I) : igas_inc + odin_sum1<real_type>(internal.n_I.data(), 0, shared->dim_n_I)));
+    state_next[5] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_I.data(), 0, shared->dim_n_I) : igas_inc + odin_sum1<real_type>(internal.n_I.data(), 0, shared->dim_n_I)));
     for (int i = 1; i <= shared->dim_gas_E_1; ++i) {
       for (int j = 1; j <= shared->dim_gas_E_2; ++j) {
         internal.gas_E[i - 1 + shared->dim_gas_E_1 * (j - 1)] = ((j == 1 ? internal.n_UE[i - 1] : internal.n_E[shared->dim_n_E_1 * (j - 1 - 1) + i - 1])) - internal.n_E[shared->dim_n_E_1 * (j - 1) + i - 1];
@@ -698,19 +706,21 @@ public:
         state_next[shared->offset_variable_R + i - 1 + shared->dim_R_1 * (j - 1)] = R[shared->dim_R_1 * (j - 1) + i - 1] + internal.dem_R[shared->dim_dem_R_1 * (j - 1) + i - 1] + internal.gas_R[shared->dim_gas_R_1 * (j - 1) + i - 1];
       }
     }
-    state_next[15] = ((shared->n_group == 16 ? internal.scarlet_fever_inc_by_group[0] / (real_type) internal.w[0] : 0));
-    state_next[16] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 1, 3) / (real_type) odin_sum1<real_type>(internal.w.data(), 1, 3) : 0));
-    state_next[17] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 3, 9) / (real_type) odin_sum1<real_type>(internal.w.data(), 3, 9) : 0));
-    state_next[18] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 9, 13) / (real_type) odin_sum1<real_type>(internal.w.data(), 9, 13) : 0));
-    state_next[19] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 13, 15) / (real_type) odin_sum1<real_type>(internal.w.data(), 13, 15) : 0));
-    state_next[20] = ((shared->n_group == 16 ? internal.scarlet_fever_inc_by_group[15] / (real_type) internal.w[15] : 0));
+    state_next[10] = odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 0, shared->dim_scarlet_fever_inc_by_group) / (real_type) odin_sum1<real_type>(internal.w.data(), 0, shared->dim_w);
+    state_next[17] = ((shared->n_group == 16 ? internal.scarlet_fever_inc_by_group[0] / (real_type) internal.w[0] : 0));
+    state_next[18] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 1, 3) / (real_type) odin_sum1<real_type>(internal.w.data(), 1, 3) : 0));
+    state_next[19] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 3, 9) / (real_type) odin_sum1<real_type>(internal.w.data(), 3, 9) : 0));
+    state_next[20] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 9, 13) / (real_type) odin_sum1<real_type>(internal.w.data(), 9, 13) : 0));
+    state_next[21] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 13, 15) / (real_type) odin_sum1<real_type>(internal.w.data(), 13, 15) : 0));
+    state_next[22] = ((shared->n_group == 16 ? internal.scarlet_fever_inc_by_group[15] / (real_type) internal.w[15] : 0));
+    state_next[4] = std::round(odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 0, shared->dim_scarlet_fever_inc_by_group) * q_F_t);
     state_next[3] = odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 0, shared->dim_scarlet_fever_inc_by_group);
-    state_next[21] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 0, 1) : 0));
-    state_next[22] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 1, 3) : 0));
-    state_next[23] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 3, 9) : 0));
-    state_next[24] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 9, 13) : 0));
-    state_next[25] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 13, 15) : 0));
-    state_next[26] = ((shared->n_group == 16 ? internal.scarlet_fever_inc_by_group[15] : 0));
+    state_next[23] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 0, 1) : 0));
+    state_next[24] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 1, 3) : 0));
+    state_next[25] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 3, 9) : 0));
+    state_next[26] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 9, 13) : 0));
+    state_next[27] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.scarlet_fever_inc_by_group.data(), 13, 15) : 0));
+    state_next[28] = ((shared->n_group == 16 ? internal.scarlet_fever_inc_by_group[15] : 0));
     for (int i = 1; i <= shared->dim_gas_A_1; ++i) {
       for (int j = 1; j <= shared->dim_gas_A_2; ++j) {
         internal.gas_A[i - 1 + shared->dim_gas_A_1 * (j - 1)] = ((j == 1 ? internal.n_UA[i - 1] : internal.n_A[shared->dim_n_A_1 * (j - 1 - 1) + i - 1])) - internal.n_A[shared->dim_n_A_1 * (j - 1) + i - 1];
@@ -727,15 +737,15 @@ public:
       }
     }
     for (int i = 1; i <= shared->dim_U; ++i) {
-      state_next[27 + i - 1] = U[i - 1] + internal.dem_U[i - 1] + internal.gas_U[i - 1];
+      state_next[29 + i - 1] = U[i - 1] + internal.dem_U[i - 1] + internal.gas_U[i - 1];
     }
-    state_next[8] = odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 0, shared->dim_gas_pharyngitis_inc_by_group) / (real_type) odin_sum1<real_type>(internal.w.data(), 0, shared->dim_w);
-    state_next[9] = ((shared->n_group == 16 ? internal.gas_pharyngitis_inc_by_group[0] / (real_type) internal.w[0] : 0));
-    state_next[10] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 1, 3) / (real_type) odin_sum1<real_type>(internal.w.data(), 1, 3) : 0));
-    state_next[11] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 3, 9) / (real_type) odin_sum1<real_type>(internal.w.data(), 3, 9) : 0));
-    state_next[12] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 9, 13) / (real_type) odin_sum1<real_type>(internal.w.data(), 9, 13) : 0));
-    state_next[13] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 13, 15) / (real_type) odin_sum1<real_type>(internal.w.data(), 13, 15) : 0));
-    state_next[14] = ((shared->n_group == 16 ? internal.gas_pharyngitis_inc_by_group[15] / (real_type) internal.w[15] : 0));
+    state_next[9] = odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 0, shared->dim_gas_pharyngitis_inc_by_group) / (real_type) odin_sum1<real_type>(internal.w.data(), 0, shared->dim_w);
+    state_next[11] = ((shared->n_group == 16 ? internal.gas_pharyngitis_inc_by_group[0] / (real_type) internal.w[0] : 0));
+    state_next[12] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 1, 3) / (real_type) odin_sum1<real_type>(internal.w.data(), 1, 3) : 0));
+    state_next[13] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 3, 9) / (real_type) odin_sum1<real_type>(internal.w.data(), 3, 9) : 0));
+    state_next[14] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 9, 13) / (real_type) odin_sum1<real_type>(internal.w.data(), 9, 13) : 0));
+    state_next[15] = ((shared->n_group == 16 ? odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 13, 15) / (real_type) odin_sum1<real_type>(internal.w.data(), 13, 15) : 0));
+    state_next[16] = ((shared->n_group == 16 ? internal.gas_pharyngitis_inc_by_group[15] / (real_type) internal.w[15] : 0));
     state_next[2] = odin_sum1<real_type>(internal.gas_pharyngitis_inc_by_group.data(), 0, shared->dim_gas_pharyngitis_inc_by_group);
     state_next[1] = ((fmodr<real_type>(step, shared->steps_per_week) == 0 ? odin_sum1<real_type>(internal.n_UE.data(), 0, shared->dim_n_UE) + odin_sum1<real_type>(internal.n_UA.data(), 0, shared->dim_n_UA) : infections_inc + odin_sum1<real_type>(internal.n_UE.data(), 0, shared->dim_n_UE) + odin_sum1<real_type>(internal.n_UA.data(), 0, shared->dim_n_UA)));
     for (int i = 1; i <= shared->dim_A_1; ++i) {
@@ -988,6 +998,7 @@ dust::pars_type<model> dust_pars<model>(cpp11::list user) {
   shared->initial_daily_gas_pharyngitis_rate_45_64 = 0;
   shared->initial_daily_gas_pharyngitis_rate_65_74 = 0;
   shared->initial_daily_gas_pharyngitis_rate_75 = 0;
+  shared->initial_daily_scarlet_fever_rate = 0;
   shared->initial_daily_scarlet_fever_rate_04 = 0;
   shared->initial_daily_scarlet_fever_rate_05_14 = 0;
   shared->initial_daily_scarlet_fever_rate_15_44 = 0;
@@ -998,6 +1009,7 @@ dust::pars_type<model> dust_pars<model>(cpp11::list user) {
   shared->initial_igas_inc = 0;
   shared->initial_infections_inc = 0;
   shared->initial_net_leavers_inc = 0;
+  shared->initial_scarlet_fever_cases = 0;
   shared->initial_scarlet_fever_inc = 0;
   shared->initial_scarlet_fever_inc_04 = 0;
   shared->initial_scarlet_fever_inc_05_14 = 0;
@@ -1053,6 +1065,9 @@ dust::pars_type<model> dust_pars<model>(cpp11::list user) {
   shared->p_I = user_get_scalar<real_type>(user, "p_I", shared->p_I, NA_REAL, NA_REAL);
   shared->p_R = user_get_scalar<real_type>(user, "p_R", shared->p_R, NA_REAL, NA_REAL);
   shared->p_S = user_get_scalar<real_type>(user, "p_S", shared->p_S, NA_REAL, NA_REAL);
+  std::array <int, 1> dim_q_F;
+  shared->q_F = user_get_array_variable<real_type, 1>(user, "q_F", shared->q_F, dim_q_F, NA_REAL, NA_REAL);
+  shared->dim_q_F = shared->q_F.size();
   shared->r_age = user_get_scalar<real_type>(user, "r_age", shared->r_age, NA_REAL, NA_REAL);
   shared->sigma = user_get_scalar<real_type>(user, "sigma", shared->sigma, NA_REAL, NA_REAL);
   shared->t0 = user_get_scalar<real_type>(user, "t0", shared->t0, NA_REAL, NA_REAL);
@@ -1263,10 +1278,10 @@ dust::pars_type<model> dust_pars<model>(cpp11::list user) {
   shared->dim_n_S = shared->dim_n_S_1 * shared->dim_n_S_2;
   shared->dim_n_Si = shared->dim_n_Si_1 * shared->dim_n_Si_2;
   shared->dim_n_Sx = shared->dim_n_Sx_1 * shared->dim_n_Sx_2;
-  shared->offset_variable_A = shared->dim_N + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
-  shared->offset_variable_N = shared->dim_U + 27;
-  shared->offset_variable_prev_A = shared->dim_N + shared->dim_U + 27;
-  shared->offset_variable_prev_R = shared->dim_N + shared->dim_U + shared->dim_prev_A + 27;
+  shared->offset_variable_A = shared->dim_N + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
+  shared->offset_variable_N = shared->dim_U + 29;
+  shared->offset_variable_prev_A = shared->dim_N + shared->dim_U + 29;
+  shared->offset_variable_prev_R = shared->dim_N + shared->dim_U + shared->dim_prev_A + 29;
   shared->omega = user_get_array_fixed<real_type, 1>(user, "omega", shared->omega, {shared->dim_omega}, NA_REAL, NA_REAL);
   shared->A0 = user_get_array_fixed<real_type, 2>(user, "A0", shared->A0, {shared->dim_A0_1, shared->dim_A0_2}, NA_REAL, NA_REAL);
   shared->E0 = user_get_array_fixed<real_type, 2>(user, "E0", shared->E0, {shared->dim_E0_1, shared->dim_E0_2}, NA_REAL, NA_REAL);
@@ -1315,11 +1330,11 @@ dust::pars_type<model> dust_pars<model>(cpp11::list user) {
     shared->initial_U[i - 1] = shared->U0[i - 1];
   }
   shared->m = user_get_array_fixed<real_type, 2>(user, "m", shared->m, {shared->dim_m_1, shared->dim_m_2}, NA_REAL, NA_REAL);
-  shared->offset_variable_E = shared->dim_A + shared->dim_N + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
-  shared->offset_variable_F = shared->dim_A + shared->dim_E + shared->dim_N + shared->dim_P + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
-  shared->offset_variable_P = shared->dim_A + shared->dim_E + shared->dim_N + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
-  shared->offset_variable_R = shared->dim_A + shared->dim_E + shared->dim_F + shared->dim_N + shared->dim_P + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
-  shared->offset_variable_S = shared->dim_A + shared->dim_E + shared->dim_N + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 27;
+  shared->offset_variable_E = shared->dim_A + shared->dim_N + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
+  shared->offset_variable_F = shared->dim_A + shared->dim_E + shared->dim_N + shared->dim_P + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
+  shared->offset_variable_P = shared->dim_A + shared->dim_E + shared->dim_N + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
+  shared->offset_variable_R = shared->dim_A + shared->dim_E + shared->dim_F + shared->dim_N + shared->dim_P + shared->dim_S + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
+  shared->offset_variable_S = shared->dim_A + shared->dim_E + shared->dim_N + shared->dim_U + shared->dim_prev_A + shared->dim_prev_R + 29;
   for (int i = 1; i <= shared->dim_r_A; ++i) {
     shared->r_A[i - 1] = shared->k_A / (real_type) shared->delta_A;
   }
@@ -1383,8 +1398,8 @@ template <>
 cpp11::sexp dust_info<model>(const dust::pars_type<model>& pars) {
   const model::internal_type internal = pars.internal;
   const std::shared_ptr<const model::shared_type> shared = pars.shared;
-  cpp11::writable::strings nms({"time", "infections_inc", "gas_pharyngitis_inc", "scarlet_fever_inc", "igas_inc", "births_inc", "net_leavers_inc", "beta_t", "daily_gas_pharyngitis_rate", "daily_gas_pharyngitis_rate_04", "daily_gas_pharyngitis_rate_05_14", "daily_gas_pharyngitis_rate_15_44", "daily_gas_pharyngitis_rate_45_64", "daily_gas_pharyngitis_rate_65_74", "daily_gas_pharyngitis_rate_75", "daily_scarlet_fever_rate_04", "daily_scarlet_fever_rate_05_14", "daily_scarlet_fever_rate_15_44", "daily_scarlet_fever_rate_45_64", "daily_scarlet_fever_rate_65_74", "daily_scarlet_fever_rate_75", "scarlet_fever_inc_04", "scarlet_fever_inc_05_14", "scarlet_fever_inc_15_44", "scarlet_fever_inc_45_64", "scarlet_fever_inc_65_74", "scarlet_fever_inc_75", "U", "N", "prev_A", "prev_R", "A", "E", "S", "P", "F", "R"});
-  cpp11::writable::list dim(37);
+  cpp11::writable::strings nms({"time", "infections_inc", "gas_pharyngitis_inc", "scarlet_fever_inc", "scarlet_fever_cases", "igas_inc", "births_inc", "net_leavers_inc", "beta_t", "daily_gas_pharyngitis_rate", "daily_scarlet_fever_rate", "daily_gas_pharyngitis_rate_04", "daily_gas_pharyngitis_rate_05_14", "daily_gas_pharyngitis_rate_15_44", "daily_gas_pharyngitis_rate_45_64", "daily_gas_pharyngitis_rate_65_74", "daily_gas_pharyngitis_rate_75", "daily_scarlet_fever_rate_04", "daily_scarlet_fever_rate_05_14", "daily_scarlet_fever_rate_15_44", "daily_scarlet_fever_rate_45_64", "daily_scarlet_fever_rate_65_74", "daily_scarlet_fever_rate_75", "scarlet_fever_inc_04", "scarlet_fever_inc_05_14", "scarlet_fever_inc_15_44", "scarlet_fever_inc_45_64", "scarlet_fever_inc_65_74", "scarlet_fever_inc_75", "U", "N", "prev_A", "prev_R", "A", "E", "S", "P", "F", "R"});
+  cpp11::writable::list dim(39);
   dim[0] = cpp11::writable::integers({1});
   dim[1] = cpp11::writable::integers({1});
   dim[2] = cpp11::writable::integers({1});
@@ -1412,18 +1427,20 @@ cpp11::sexp dust_info<model>(const dust::pars_type<model>& pars) {
   dim[24] = cpp11::writable::integers({1});
   dim[25] = cpp11::writable::integers({1});
   dim[26] = cpp11::writable::integers({1});
-  dim[27] = cpp11::writable::integers({shared->dim_U});
-  dim[28] = cpp11::writable::integers({shared->dim_N});
-  dim[29] = cpp11::writable::integers({shared->dim_prev_A});
-  dim[30] = cpp11::writable::integers({shared->dim_prev_R});
-  dim[31] = cpp11::writable::integers({shared->dim_A_1, shared->dim_A_2});
-  dim[32] = cpp11::writable::integers({shared->dim_E_1, shared->dim_E_2});
-  dim[33] = cpp11::writable::integers({shared->dim_S_1, shared->dim_S_2});
-  dim[34] = cpp11::writable::integers({shared->dim_P_1, shared->dim_P_2});
-  dim[35] = cpp11::writable::integers({shared->dim_F_1, shared->dim_F_2});
-  dim[36] = cpp11::writable::integers({shared->dim_R_1, shared->dim_R_2});
+  dim[27] = cpp11::writable::integers({1});
+  dim[28] = cpp11::writable::integers({1});
+  dim[29] = cpp11::writable::integers({shared->dim_U});
+  dim[30] = cpp11::writable::integers({shared->dim_N});
+  dim[31] = cpp11::writable::integers({shared->dim_prev_A});
+  dim[32] = cpp11::writable::integers({shared->dim_prev_R});
+  dim[33] = cpp11::writable::integers({shared->dim_A_1, shared->dim_A_2});
+  dim[34] = cpp11::writable::integers({shared->dim_E_1, shared->dim_E_2});
+  dim[35] = cpp11::writable::integers({shared->dim_S_1, shared->dim_S_2});
+  dim[36] = cpp11::writable::integers({shared->dim_P_1, shared->dim_P_2});
+  dim[37] = cpp11::writable::integers({shared->dim_F_1, shared->dim_F_2});
+  dim[38] = cpp11::writable::integers({shared->dim_R_1, shared->dim_R_2});
   dim.names() = nms;
-  cpp11::writable::list index(37);
+  cpp11::writable::list index(39);
   index[0] = cpp11::writable::integers({1});
   index[1] = cpp11::writable::integers({2});
   index[2] = cpp11::writable::integers({3});
@@ -1451,16 +1468,18 @@ cpp11::sexp dust_info<model>(const dust::pars_type<model>& pars) {
   index[24] = cpp11::writable::integers({25});
   index[25] = cpp11::writable::integers({26});
   index[26] = cpp11::writable::integers({27});
-  index[27] = integer_sequence(28, shared->dim_U);
-  index[28] = integer_sequence(shared->offset_variable_N + 1, shared->dim_N);
-  index[29] = integer_sequence(shared->offset_variable_prev_A + 1, shared->dim_prev_A);
-  index[30] = integer_sequence(shared->offset_variable_prev_R + 1, shared->dim_prev_R);
-  index[31] = integer_sequence(shared->offset_variable_A + 1, shared->dim_A);
-  index[32] = integer_sequence(shared->offset_variable_E + 1, shared->dim_E);
-  index[33] = integer_sequence(shared->offset_variable_S + 1, shared->dim_S);
-  index[34] = integer_sequence(shared->offset_variable_P + 1, shared->dim_P);
-  index[35] = integer_sequence(shared->offset_variable_F + 1, shared->dim_F);
-  index[36] = integer_sequence(shared->offset_variable_R + 1, shared->dim_R);
+  index[27] = cpp11::writable::integers({28});
+  index[28] = cpp11::writable::integers({29});
+  index[29] = integer_sequence(30, shared->dim_U);
+  index[30] = integer_sequence(shared->offset_variable_N + 1, shared->dim_N);
+  index[31] = integer_sequence(shared->offset_variable_prev_A + 1, shared->dim_prev_A);
+  index[32] = integer_sequence(shared->offset_variable_prev_R + 1, shared->dim_prev_R);
+  index[33] = integer_sequence(shared->offset_variable_A + 1, shared->dim_A);
+  index[34] = integer_sequence(shared->offset_variable_E + 1, shared->dim_E);
+  index[35] = integer_sequence(shared->offset_variable_S + 1, shared->dim_S);
+  index[36] = integer_sequence(shared->offset_variable_P + 1, shared->dim_P);
+  index[37] = integer_sequence(shared->offset_variable_F + 1, shared->dim_F);
+  index[38] = integer_sequence(shared->offset_variable_R + 1, shared->dim_R);
   index.names() = nms;
   size_t len = shared->offset_variable_R + shared->dim_R;
   using namespace cpp11::literals;